api/requirements.txt
aniso8601==3.0.2
autopep8==1.5.1
cachelib==0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==6.7
cryptography==2.9.2
dnspython==1.16.0
eventlet==0.25.1
Flask==1.0.2
Flask-Compress==1.4.0
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.6
Flask-Session==0.3.2
greenlet==0.4.15
gunicorn==20.0.4
http-ece==1.1.0
idna==2.9
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
Pillow==6.2.0
py-vapid==1.7.0
pycodestyle==2.5.0
pycparser==2.20
PyMySQL==0.9.3
pysodium==0.7.0.post0
python-dateutil==2.7.3
pytz==2018.5
pywebpush==1.11.0
redis==3.5.2
requests==2.23.0
six==1.11.0
urllib3==1.25.8
Werkzeug==0.15.5
